Refactor product options view by extracting logic to utils
