+++ src/app/components/productDetails/ProductDetailsComponent.js
@@ -12,36 +12,28 @@ const ProductDetailsItemComponent = ({ options }) =>
   </div>;
 
 ProductDetailsItemComponent.propTypes = {
+  options: PropTypes.object.isRequired,
   details: PropTypes.object.isRequired,
 };
 
-const ProductDetailsComponent = ({ product, productOptions, productParsedToHtml }) => {
-  const details = [
-    { key: 'Id', value: product.id },
-    { key: 'City', value: product.productCity },
-    { key: 'Category', value: product.category },
-    { key: 'SubCategory', value: product.subCategory },
-  ];
-  const productDetails = [...productOptions, ...details];
-  return (
-    <div className="ProductDetailsComponentBlock">
-      <div className="ProductDetailsComponent">
-        <div className="ProductDetailsComponent-detailsBlock">
-          <div className="ProductDetailsComponent-details">
-          {_.map(productDetails, (options, key) =>
+const ProductDetailsComponent = ({ productOptions, productParsedToHtml }) => (
+  <div className="ProductDetailsComponentBlock">
+    <div className="ProductDetailsComponent">
+      <div className="ProductDetailsComponent-detailsBlock">
+        <div className="ProductDetailsComponent-details">
+          {_.map(productOptions, (options, key) =>
             <div className="ProductDetailsComponent-details-item" key={key}>
               <ProductDetailsItemComponent options={options} />
             </div>
           )}
-          </div>
-        </div>
-        <div className="ProductDetailsComponent-description">
-          <p dangerouslySetInnerHTML={{ __html: productParsedToHtml }} />
         </div>
       </div>
+      <div className="ProductDetailsComponent-description">
+        <p dangerouslySetInnerHTML={{ __html: productParsedToHtml }} />
+      </div>
     </div>
-  );
-};
+  </div>
+);
 
 ProductDetailsComponent.propTypes = {
   product: PropTypes.object.isRequired,
