+++ src/app/utils/index.js
@@ -43,3 +43,13 @@ export const mapProductImagesToAmazonUrl = images => _.chain(images)
   .filter(item => item.filename)
   .map(item => `https://shoptemplate.s3-eu-west-1.amazonaws.com/${item.filename}`)
   .value();
+
+export const mapProductOptions = (product, productOptions) => {
+  const details = [
+    { key: 'Id', value: product.id },
+    { key: 'City', value: product.productCity },
+    { key: 'Category', value: product.category },
+    { key: 'SubCategory', value: product.subCategory },
+  ];
+  return [...productOptions, ...details];
+};
